remove tablaTemp

+++ src/java/controllers/EstudioController.java
@@ -320,7 +320,7 @@ public class EstudioController {
     @ResponseBody
     @RequestMapping(value = "/estudio/addTablaCnaGeneral/{valor:.+}")
     public ModelAndView addTablaCnaGeneral(@RequestBody TablaCnaGeneralPK cnaGeneralPK,
-            @PathVariable Double valor, ModelMap map) {
+            @PathVariable Float valor, ModelMap map) {
         try {
             TablaCnaGeneral cnaGeneral = new TablaCnaGeneral();
             cnaGeneral.setTablaCnaGeneralPK(cnaGeneralPK);
@@ -340,7 +340,7 @@ public class EstudioController {
     @ResponseBody
     @RequestMapping(value = "/estudio/editTablaCnaGeneral/{valor:.+}")
     public ModelAndView editTablaCnaGeneral(@RequestBody TablaCnaGeneralPK cnaGeneralPK,
-            @PathVariable Double valor, ModelMap map) {
+            @PathVariable Float valor, ModelMap map) {
         try {
             TablaCnaGeneral cnaGeneral = tablaCnaGeneralRepo.findOne(cnaGeneralPK);
             cnaGeneral.setValor(valor);
